fix(gpu): make gather_kernel int-indexed + right name/dispatch — clears ALL hollow overrides (#775)
The last 4 hollow overrides were the gather cluster (Gather/TensorGather/TensorTake/MaskedSelect).
Root cause was consistent, not per-caller as first feared: gather_kernel read indices as (int)float
from a FLOAT buffer, but 8 of 9 backend.Gather callers upload INT bit-patterns via AllocateIntBuffer
(only my plain Gather uploaded floats) — and the "gather" vs registered "gather_kernel" name bug meant
none of them ever launched (silent CPU fallback). Fix coherently: kernel reads __global const int*
indices; backend.Gather uses "gather_kernel" + Execute1D(numIndices) (it is a 1D axis-0 row gather, not
a 2D featureSize x numIndices grid); switch plain Gather to AllocateIntBuffer so all 9 callers agree on
int encoding. My earlier rename-only attempt broke 3 ops precisely because it left this encoding
mismatch — fixing the encoding too makes all 4 pass.

Probe-verified: hollow overrides 10 -> 0 (ALL cleared), CPU-fallback drops by 4 more, full suite green
(931, no collateral). The hollow-override floor is now 0 — any future silent kernel-not-found fallback
fails the build.

+++ src/AiDotNet.Tensors/Engines/DirectGpu/OpenCL/Kernels/NeuralNetKernels.cs
@@ -816,7 +816,7 @@ __kernel void fma_kernel(
 // Gather operation
 __kernel void gather_kernel(
     __global const float* source,
-    __global const float* indices,
+    __global const int* indices,
     __global float* output,
     const int numIndices,
     const int featureSize)
@@ -824,7 +824,7 @@ __kernel void gather_kernel(
     const int idx = get_global_id(0);
     if (idx >= numIndices) return;
 
-    int index = (int)indices[idx];
+    int index = indices[idx];
     for (int f = 0; f < featureSize; f++) {
         output[idx * featureSize + f] = source[index * featureSize + f];
     }
